routes: migrate staff + mcp domains (11 routes)
Move to:
- src/server/routes/staff.ts (7): /api/staff CRUD + /:id/wake + /:id/sessions
- src/server/routes/mcp.ts (4): /api/mcp-servers, /api/internal/mcp-call, /api/internal/mcp-describe

Net: server.ts -338 lines (6423 → 6085). No behaviour change.
