+++ src/server/routes/dispatcher.ts
@@ -25,6 +25,8 @@ import { rolesRoutes } from "./roles.js";
 import { toolsRoutes } from "./tools.js";
 import { skillsRoutes } from "./skills.js";
 import { workflowsRoutes } from "./workflows.js";
+import { staffRoutes } from "./staff.js";
+import { mcpRoutes } from "./mcp.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -41,6 +43,8 @@ const allRoutes: ReadonlyArray<Route> = [
 	...toolsRoutes,
 	...skillsRoutes,
 	...workflowsRoutes,
+	...staffRoutes,
+	...mcpRoutes,
 ];
 
 /**
