+++ src/server/server.ts
@@ -26,8 +26,7 @@ import { ToolManager } from "./agent/tool-manager.js";
 
 import { getPromptSections, initPromptDirs, loadPersistedPromptSections } from "./agent/system-prompt.js";
 import { recordElapsed } from "./agent/profiling.js";
-import { resolveGrantPolicy } from "./agent/tool-activation.js";
-import { parseMcpToolName } from "./mcp/mcp-meta.js";
+
 import { initSkillSidecarDir } from "./skills/skill-sidecar.js";
 import { buildActivationHeader } from "./skills/skill-manifest.js";
 import type { TaskState } from "./agent/task-store.js";
@@ -1057,11 +1056,11 @@ async function handleApiRoute(
 	toolManager: ToolManager,
 	projectContextManager: ProjectContextManager,
 	bgProcessManager: BgProcessManager,
-	staffManager: StaffManager,
+	_staffManager: StaffManager,
 	verificationHarness: VerificationHarness,
 	_preferencesStore: PreferencesStore,
 	projectConfigStore: ProjectConfigStore,
-	groupPolicyStore: ToolGroupPolicyStore,
+	_groupPolicyStore: ToolGroupPolicyStore,
 	broadcastToGoal: (goalId: string, event: any) => void,
 	broadcastToAll: (event: any) => void,
 	sandboxManager: SandboxManager | null,
@@ -5655,343 +5654,6 @@ async function handleApiRoute(
 		const body = await readBody(req);
 		reviewAnnotationStore.setSubmitted(sessionId, !!body?.submitted);
 		json({ ok: true });
-		return;
-	}
-
-	// ── Staff endpoints ────────────────────────────────────────────
-
-	// GET /api/staff
-	if (url.pathname === "/api/staff" && req.method === "GET") {
-		const projectId = url.searchParams.get("projectId") || undefined;
-		const list = staffManager.listStaff(projectId).map(s => {
-			const sandboxed = s.projectId ? (projectContextManager.getOrCreate(s.projectId)?.projectConfigStore.get("sandbox") === "docker") : false;
-			return { ...s, sandboxed };
-		});
-		json({ staff: list });
-		return;
-	}
-
-	// POST /api/staff
-	if (url.pathname === "/api/staff" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body?.name || typeof body.name !== "string") {
-			json({ error: "Missing name" }, 400);
-			return;
-		}
-		if (!body?.systemPrompt || typeof body.systemPrompt !== "string") {
-			json({ error: "Missing systemPrompt" }, 400);
-			return;
-		}
-		const cwd = body.cwd || config.defaultCwd;
-		const resolved = resolveProjectForRequest(projectRegistry, projectContextManager, { projectId: body.projectId, cwd });
-		if (!resolved.ok) {
-			json({ error: resolved.error }, resolved.status);
-			return;
-		}
-		const projectId = resolved.projectId;
-		try {
-			const staff = await staffManager.createStaff(
-				body.name,
-				body.description || "",
-				body.systemPrompt,
-				cwd,
-				sessionManager,
-				{ triggers: body.triggers, roleId: body.roleId, projectId, sandboxed: body.sandboxed },
-			);
-			json(staff, 201);
-		} catch (err: any) {
-			console.error("[server] Failed to create staff agent:", err);
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// Routes with staff :id parameter
-	const staffMatch = url.pathname.match(/^\/api\/staff\/([^/]+)$/);
-	if (staffMatch) {
-		const id = staffMatch[1];
-
-		if (req.method === "GET") {
-			const staff = staffManager.getStaff(id);
-			if (!staff) { json({ error: "Staff agent not found" }, 404); return; }
-			const sandboxed = staff.projectId ? (projectContextManager.getOrCreate(staff.projectId)?.projectConfigStore.get("sandbox") === "docker") : false;
-			json({ ...staff, sandboxed });
-			return;
-		}
-
-		if (req.method === "PUT") {
-			const body = await readBody(req);
-			if (!body) { json({ error: "Missing body" }, 400); return; }
-			const ok = staffManager.updateStaff(id, {
-				name: body.name,
-				description: body.description,
-				systemPrompt: body.systemPrompt,
-				cwd: body.cwd,
-				state: body.state,
-				triggers: body.triggers,
-				memory: body.memory,
-				roleId: body.roleId,
-			});
-			if (!ok) { json({ error: "Staff agent not found" }, 404); return; }
-			json(staffManager.getStaff(id));
-			return;
-		}
-
-		if (req.method === "DELETE") {
-			const ok = await staffManager.deleteStaff(id, sessionManager);
-			if (!ok) { json({ error: "Staff agent not found" }, 404); return; }
-			json({ ok: true });
-			return;
-		}
-	}
-
-	// POST /api/staff/:id/wake — manually trigger a wake cycle
-	const staffWakeMatch = url.pathname.match(/^\/api\/staff\/([^/]+)\/wake$/);
-	if (staffWakeMatch && req.method === "POST") {
-		const id = staffWakeMatch[1];
-		const staff = staffManager.getStaff(id);
-		if (!staff) { json({ error: "Staff agent not found" }, 404); return; }
-		const body = await readBody(req);
-		try {
-			const sessionId = await staffManager.wake(id, body?.prompt, sessionManager);
-			json({ sessionId }, 201);
-		} catch (err) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// GET /api/staff/:id/sessions — DEPRECATED (staff agents have a single permanent session)
-	const staffSessionsMatch = url.pathname.match(/^\/api\/staff\/([^/]+)\/sessions$/);
-	if (staffSessionsMatch && req.method === "GET") {
-		json({ error: "Deprecated. Staff agents have a single permanent session. Use GET /api/staff/:id." }, 410);
-		return;
-	}
-
-	// GET /api/mcp-servers
-	if (url.pathname === "/api/mcp-servers" && req.method === "GET") {
-		const mcpManager = sessionManager.getMcpManager();
-		if (!mcpManager) {
-			json([]);
-			return;
-		}
-		const statuses = mcpManager.getServerStatuses();
-		const toolInfos = mcpManager.getToolInfos();
-		const result = statuses.map(s => ({
-			...s,
-			tools: toolInfos.filter(t => t.serverName === s.name).map(t => {
-				const parsed = parseMcpToolName(t.name);
-				return {
-					name: t.name,
-					description: t.description,
-					subNamespace: parsed?.sub,
-					op: parsed?.op ?? t.mcpToolName,
-				};
-			}),
-		}));
-		json(result);
-		return;
-	}
-
-	// POST /api/mcp-servers/:name/restart
-	const mcpRestartMatch = url.pathname.match(/^\/api\/mcp-servers\/([^/]+)\/restart$/);
-	if (mcpRestartMatch && req.method === "POST") {
-		const mcpManager = sessionManager.getMcpManager();
-		if (!mcpManager) {
-			json({ error: "MCP not initialized" }, 500);
-			return;
-		}
-		const serverName = decodeURIComponent(mcpRestartMatch[1]);
-		let statuses = mcpManager.getServerStatuses();
-		let existing = statuses.find(s => s.name === serverName);
-		if (!existing || !existing.config) {
-			// Re-discover servers in case config was added after startup
-			const discovered = mcpManager.discoverServers();
-			if (!discovered[serverName]) {
-				json({ error: `MCP server "${serverName}" not found` }, 404);
-				return;
-			}
-			// Connect the newly discovered server
-			await mcpManager.connectServer(serverName, discovered[serverName]);
-		} else {
-			await mcpManager.disconnectServer(serverName);
-			// Re-discover to pick up any config changes from disk
-			const refreshed = mcpManager.discoverServers();
-			const config = refreshed[serverName] || existing.config;
-			await mcpManager.connectServer(serverName, config);
-		}
-		// Re-register MCP tools with ToolManager
-		if (toolManager) {
-			toolManager.removeExternalTools("mcp__");
-			const infos = mcpManager.getToolInfos();
-			toolManager.registerExternalTools(infos.map(info => ({
-				name: info.name,
-				description: info.description,
-				summary: info.description,
-				group: info.group,
-				docs: info.docs,
-				provider: { type: 'mcp' as const, server: info.serverName, mcpTool: info.mcpToolName },
-			})));
-		}
-		const updated = mcpManager.getServerStatuses().find(s => s.name === serverName);
-		json({ ok: true, ...updated });
-		return;
-	}
-
-	// POST /api/internal/mcp-call
-	if (url.pathname === "/api/internal/mcp-call" && req.method === "POST") {
-		const mcpManager = sessionManager.getMcpManager();
-		if (!mcpManager) {
-			json({ error: "MCP not initialized" }, 500);
-			return;
-		}
-		let parsedToolForError: string | undefined;
-		try {
-			const body = await new Promise<string>((resolve) => {
-				let data = "";
-				req.on("data", (chunk: Buffer) => data += chunk.toString());
-				req.on("end", () => resolve(data));
-			});
-			const { tool, args } = JSON.parse(body);
-			parsedToolForError = typeof tool === "string" ? tool : undefined;
-			if (!tool) {
-				json({ error: "Missing 'tool' field" }, 400);
-				return;
-			}
-
-			// Enforce allowedTools for the calling session.
-			// This endpoint is internal — only MCP proxy extensions should call it.
-			// Require session ID header to prevent direct curl bypass.
-			const mcpSessionId = req.headers["x-bobbit-session-id"] as string | undefined;
-			if (!mcpSessionId) {
-				json({ error: "Missing X-Bobbit-Session-Id header" }, 403);
-				return;
-			}
-			// Verify the session exists (live or persisted).
-			const mcpSession = sessionManager.getSession(mcpSessionId);
-			const persistedSession = mcpSession ? null : (
-				// Search across all project stores for persisted session
-				projectContextManager.getContextForSession(mcpSessionId)?.sessionStore.get(mcpSessionId)
-				?? null
-			);
-			if (!mcpSession && !persistedSession) {
-				json({ error: `Session "${mcpSessionId}" not found` }, 403);
-				return;
-			}
-			// Enforce allowedTools for non-MCP tools on live sessions.
-			// MCP tools (mcp__*) are dynamically discovered and governed by the
-			// grant policy system — they may not appear in the session's static
-			// allowedTools list, so we skip the check for them.
-			const toolStr = tool as string;
-			if (!toolStr.startsWith("mcp__") && mcpSession?.allowedTools && mcpSession.allowedTools.length > 0) {
-				if (!mcpSession.allowedTools.some((t: string) => t.toLowerCase() === toolStr.toLowerCase())) {
-					json({ error: `Tool "${tool}" is not allowed for this session` }, 403);
-					return;
-				}
-			}
-
-			// Layer B per-op never-policy enforcement (§4.3). Resolves the per-op
-			// policy for `mcp__<server>__<op>` even though the model only sees the
-			// aggregated `mcp_<server>` meta-tool — keeps `never` denials honoured
-			// after the meta-tool is granted wholesale.
-			if (toolStr.startsWith("mcp__")) {
-				const roleName = mcpSession?.role ?? (persistedSession as any)?.role;
-				const role = roleName ? roleManager.getRole(roleName) : undefined;
-				const parsed = parseMcpToolName(toolStr);
-				const opGroup = parsed?.server ? `MCP: ${parsed.server}` : undefined;
-				const policy = resolveGrantPolicy(toolStr, opGroup, role, toolManager, groupPolicyStore);
-				if (policy === "never") {
-					json({ error: `tool ${toolStr} denied by policy`, tool: toolStr, reason: "policy=never" }, 403);
-					return;
-				}
-			}
-
-			const result = await mcpManager.callTool(tool, args || {});
-			json(result);
-		} catch (err) {
-			const e = err as Error;
-			console.error(`[mcp] Tool call failed:`, e.stack || e);
-			// Parse `mcp__<server>__<op>` into structured fields (§5.4).
-			let parsedServer: string | undefined;
-			let parsedOperation: string | undefined;
-			if (parsedToolForError && parsedToolForError.startsWith("mcp__")) {
-				const parsedErr = parseMcpToolName(parsedToolForError);
-				if (parsedErr) {
-					parsedServer = parsedErr.server;
-					parsedOperation = parsedErr.sub ? `${parsedErr.sub}__${parsedErr.op}` : parsedErr.op;
-				}
-			}
-			json({ error: e.message, server: parsedServer, operation: parsedOperation, stack: e.stack }, 500);
-		}
-		return;
-	}
-
-	// POST /api/internal/mcp-describe — discovery endpoint for the `mcp_describe` tool (§3.3)
-	if (url.pathname === "/api/internal/mcp-describe" && req.method === "POST") {
-		const mcpManager = sessionManager.getMcpManager();
-		if (!mcpManager) {
-			json({ error: "MCP not initialized" }, 500);
-			return;
-		}
-		try {
-			const body = await new Promise<string>((resolve) => {
-				let data = "";
-				req.on("data", (chunk: Buffer) => data += chunk.toString());
-				req.on("end", () => resolve(data));
-			});
-			const parsed = JSON.parse(body || "{}");
-			const server: string | undefined = parsed?.server;
-			const operation: string | undefined = parsed?.operation;
-			if (!server || typeof server !== "string") {
-				json({ error: "Missing 'server' field" }, 400);
-				return;
-			}
-
-			const describeSessionId = req.headers["x-bobbit-session-id"] as string | undefined;
-			if (!describeSessionId) {
-				json({ error: "Missing X-Bobbit-Session-Id header" }, 403);
-				return;
-			}
-			const liveSession = sessionManager.getSession(describeSessionId);
-			const persistedSession = liveSession ? null : (
-				projectContextManager.getContextForSession(describeSessionId)?.sessionStore.get(describeSessionId)
-				?? null
-			);
-			if (!liveSession && !persistedSession) {
-				json({ error: `Session "${describeSessionId}" not found` }, 403);
-				return;
-			}
-
-			const statuses = mcpManager.getServerStatuses();
-			const status = statuses.find(s => s.name === server);
-			if (!status || status.status !== "connected") {
-				const reason = status?.error ?? (status ? status.status : "unknown server");
-				json({ error: `server ${server} not connected: ${reason}` }, 503);
-				return;
-			}
-
-			const infos = mcpManager.getToolInfos().filter(i => i.serverName === server);
-			if (operation) {
-				const match = infos.find(i => i.mcpToolName === operation);
-				if (!match) {
-					json({ error: "operation not found" }, 404);
-					return;
-				}
-				json({ tool: { name: match.mcpToolName, description: match.description, inputSchema: match.inputSchema } });
-				return;
-			}
-			json({
-				tools: infos.map(i => ({
-					name: i.mcpToolName,
-					description: i.description,
-					inputSchema: i.inputSchema,
-				})),
-			});
-		} catch (err) {
-			const e = err as Error;
-			console.error(`[mcp] Describe failed:`, e.stack || e);
-			json({ error: e.message, stack: e.stack }, 500);
-		}
 		return;
 	}
